requirements.txt
# ==============================
# Agentic AI Requirements (Runtime + Development)
# ==============================

# Agentic / LLM framework
langgraph-checkpoint-sqlite
langgraph
langgraph-prebuilt
langgraph-supervisor
langchain
langchain-community
langchain-experimental
langchain-tavily
langchain-core
langsmith
langchain-openai
langchain-anthropic
langchain-google-genai
langchain-groq
langchain-ollama
langchain-huggingface

# Model providers
openai
anthropic
google-generativeai
groq
litellm

# Hugging Face models and tokenizers
transformers
torch
accelerate
sentence-transformers
tiktoken

# Vector db
chromadb
qdrant-client

# API / Web
fastapi
pydantic
pydantic-settings
uvicorn[standard]
sse-starlette
python-multipart
orjson

# Database
# SQLite (built-in with Python)

# Cache
redis

# Observability / tracing
opentelemetry-sdk

# Config utils
python-dotenv

# Search
duckduckgo-search
ddgs

# Data visualization
matplotlib

# ==============================
# Development Dependencies
# ==============================

# Formatting
black
isort

# Linting
ruff

# Type checking
mypy

# Testing
pytest
pytest-asyncio
pytest-cov
httpx

# Security
bandit

# Git hooks
pre-commit